Replace product display with conditional rendering in product container
